Stop using soon-to-be-deprecated Spring Security API
Spring Security 6.4's Kotlin DSL deprecates the `authorizeRequests` block and
suggests using `authorizeHttpRequests` instead. Make that change so that the
upgrade will succeed rather than failing to compile with a deprecation warning.

The upgrade requires a workaround for a missing feature in the Kotlin DSL; when
they've addressed that, we can remove the workaround.