SW-6293 Fix seed withdrawals of weight-based accessions
If a user withdraws a seed quantity from a weight-based accession, we use the
subset weight and count to calculate how much weight to subtract from the
accession. This can be off by a bit due to the limited precision of subset weight
measurements, so we had a fudge factor to try to detect if the user was trying to
withdraw all the remaining seeds and set the remaining weight to zero if so.

Unfortunately, the fudge factor wasn't enough for accessions measured in
kilograms or pounds where the weight of a single seed can round down to zero.

Replace the fudge factor with a more direct approach: see if the withdrawal
quantity is the same as the estimated seed count.