Upgrade to ktfmt 0.53
Move to the latest version of ktfmt. It includes a change that's not compatible
with the Spotless plugin version we were using, so upgrade that as well;
unfortunately the only version that fixes the compatibility issue is a beta
release, but none of the breaking changes in the beta appear to be relevant to
our code base.