SW-6312 Use ASCII apostrophe in Côte d'Ivoire
The country list in the all-variables sheet uses an ASCII apostrophe in the name
of the country, but Terraware was using a right single quotation mark, meaning
it wasn't working to select that country name for the country variable.

Update the Terraware country list to use the ASCII apostrophe to match what's in
the spreadsheet.