SW-6251 Introduce exterior monitoring plots
Add the concept of an "exterior" monitoring plot, which is associated with a
planting site but not with a subzone.

This change updates the data model and adds internal support for reading
exterior plots, but there isn't yet any way to create an exterior plot outside
the test suite.

Other upcoming changes will make it possible for a monitoring plot to become
exterior if the site map is edited such that the plot no longer falls within
any subzone, or if the plot is created for an ad-hoc observation.