SW-6295 Fix deleting projects with submitted reports (#2645)
If you delete a project for which you had previously submitted reports, we keep
the reports around but set their project IDs to null. This caused problems if
you also had org-level reports for the same year and quarter because we have a
constraint to allow only one org-level report per quarter.

The constraint was assuming any report with a null project ID was an org-level
report, meaning that when the project ID was set to null on the project-level
reports, they conflicted with the actual org-level reports.

Fix it by making the constraint verify that both the project ID and the project
name are null. For reports from deleted projects, the project name is retained,
but both the ID and name are null for org-level reports.