SW-6280 Track observation site versions
Record which version of a planting site is used at observation start time.
Knowing the site history ID for each observation will allow the app to show the
correct version of a site map if the map is edited between observations; the
history IDs will be exposed in the API in subsequent code changes.