SW-6344 Make site map editing behavior selectable (#2679)
While the new site map editing features are under development, admins still need
to be able to update existing sites with the original map editing behavior. Add a
drop-down menu (currently with just one option) to eventually allow the admin to
select which version of the map editing feature they want to use.

Once the new version is fully released, we'll get rid of the original one and the
new UI element can go away.