Record lower permission check in helper functions (#2682)
Change `isGlobalReader` and `isGlobalWriter` in `IndividualUser` so they don't
record a check for the super-admin global role in the permission check history,
just the check for the less-powerful role that also requires an organization to
have an appropriate internal tag. This will let the permission inversion code
detect cases where we do one of those permission checks and later do something
that requires the super-admin role.