SW-6368 Allow withdrawals from backdated batches
If a user wants to record a historical nursery batch, they create it with an added
date in the past and a withdrawal date in the past. The system wasn't treating the
withdrawal as a quantity change in that case because it assumed that the seedling
quantities entered by the user when they created the batch were the quantities at
the time the batch was being created, and thus that older withdrawals shouldn't
affect the more recent quantity values supplied by the user.

Change batch creation such that if a batch's added date is in the past (based on
the nursery's time zone), we treat its quantities as having been observed at the
start of that day. Backdated withdrawals will thus be properly deducted from the
batch's quantities.