SW-6368 Fix withdrawal backfill migration (#2713)
The backfill migration from commit b2cf7e7 wasn't handling cases where a user
retried a full-batch withdrawal after seeing the quantity not change the first
time they withdrew. There were no cases of this in the production database but
there were some in the staging database, which caused the migration to fail
because it tried to insert duplicate quantity history entries.