SW-6344 Remove unused monitoring plot edit calculation
The planting site edit calculator was returning a list of monitoring plots to
remove, but that list was never actually used for anything; plots are deleted by
different logic in `PlantingSiteStore`, and the upcoming changes to planting
site editing will deal with plots in a completely different way. Remove the
unused field.