Migration to convert remaining 25-meter plots
Add a migration to the admin UI to replace the remaining permanent clusters of
25-meter monitoring plots with 30-meter plots.

This will only affect sites that haven't had observations since the introduction
of 30-meter plots, since the conversion also happens when an observation starts.

To account for very large test sites where floating-point inaccuracy becomes more
significant in coordinate calculations and can cause replacement plots to appear
to fall slightly outside the boundary of the old cluster, relax the "is this plot
located inside the search boundary?" check slightly, from 99.99% coverage to 99.9%.

Once this has been run, we can start to remove support for multi-plot clusters,
which will mean we don't need to account for them as we add support for more
flexible site map editing.