SW-6416 Read zone settings from shapefiles (#2744)
We want to allow setting the inputs to the plot count calculations at the same
time the site map is updated so that any adjustments to plots will take both
the map changes and the plot count changes into account.

Update the shapefile importer to require the error margin and variance values
to be supplied as planting subzone properties. The Student's t value may also
be overridden in a property but isn't required. As before, the shapefile can
also directly set the permanent and temporary plot counts, but the calculation
inputs are still required even if the counts are overridden.