Add zipProportionally function (#2752)
The new planting site editing code will support changing planting zone boundaries
and will assign monitoring plots to the newly-added area based on how big a
fraction of the zone's total area is newly added. This will require constructing
a list of monitoring plots that come from the old and new areas in the right
proportions.

The algorithm for proportional allocation isn't specific to monitoring plots, so
add a general-purpose function that zips together elements from two sources with
a specified ratio between the two. The term "zip" is already used in the Kotlin
standard library for the concept of interleaving elements from two lists.