SW-6444 Track zone boundary modification time (#2758)
When the user is selecting subzones to observe, we want to include an indication
of when the zone boundary was last edited. Add a boundary modification time (and
user ID) to the planting zones table and bump it whenever a zone edit causes an
actual change to the boundary geometry. Edits that affect things other than the
boundary, e.g., changes to plot counts, don't affect this new timestamp.

The timestamp is included in the planting zone API payload and is also available
via the search API.