SW-6462 Populate default text for unmodified sections
If a section variable's definition hasn't changed since the previous manifest
for the same document template, we reuse the existing variable ID. However, the
default section text is per-manifest data (changing it doesn't prevent existing
variables from being reused) so it needs to be inserted into the database with
the new variable manifest ID even if it's the same as in the previous manifest.

This was causing newly-created documents to be missing default text if they used
a template whose manifest had been uploaded more than once.