SW-6455 Implemented Biomass Species tables (#2770)
Herbaceous and tree species are required to have `isThreatened` and `isInvasive` attributes within a biomass observation consistently. 

This PR introduces two tables, to track these attributes, and sets up foreign keys so that any referenced herbaceous and tree species must be in those tables. 

The API and store payloads are updated accordingly.