SW-6344 Add behavior field to PlantingSiteEdit (#2785)
Some of the downstream behavior of planting site edits, e.g., the effect on active
observations, will need to differ depending on whether an edit uses the original
restricted editing feature or the upcoming more flexible editing style. In
preparation, add a field to `PlantingSiteEdit` to indicate which behavior to use.

In this change, there is only one possible value; an additional one will be added
later.