Used weighted standard deviation for mortality rate (#2797)
Our previous mortality rate calculations treat each plot mortality rate as one data point. This is incorrect because our mortality rate in larger sampling sizes (subzones, zones, sites) actually calculates the value, by adding up all the dead and live plants found in the plots. This is equivalent to a weighted mean. 

To correct the calculations, we need to use the variance of mortality rate weighted by the number of plants found in that plot instead. So a plot with X number of plants has a weight of X when calculating this result.

Another way to think of this is, if a plot has X number of plants and Y mortality rate, we are counting this as X number of data points, each with a value Y. So the sum of difference squared and the variance calculation will now correctly factored in the number of plants within a plot.

https://docs.google.com/spreadsheets/d/17-9VlElRgnXU7rnCPsesw6bM5c1Du7Z_ZIvSq8yJPbI/edit?usp=sharing