SW-6321 Allow unplanted subzones to be observed
Previously, we only allowed observation of plots in subzones where plants had
previously been delivered from nurseries, so as to avoid asking people to
survey parts of a planting site they hadn't started working on yet. Now that
users are required to list the subzones they want to observe, this restriction
isn't useful, and it forces users who aren't tracking plants in nurseries to
create fake nursery batches just to be able to schedule observations.

Remove all the "is the subzone planted?" checks from the plot selection logic,
and instead rely solely on the observation's list of requested subzones.