SW-6344 Eject monitoring plots outside of site after edit (#2790)
Previously, we didn't have the concept of a monitoring plot that wasn't associated
with a subzone, so if a map edit shrank the planting site such that a monitoring
plot was no longer inside the site boundary, we kept it associated with its old
subzone.

But now we have external monitoring plots, so we can more correctly model this
situation by "ejecting" a monitoring plot from its subzone if it's outside the
site. (The current "restricted" site editing feature doesn't allow moving a
plot from one subzone to another.)

Move the "is the plot outside the site now?" calculation to the edit calculator,
which now adds `MonitoringPlotEdit.Eject` objects to the affected subzones.