SW-6344 Reuse exterior plots when zone expands (#2796)
If a planting site edit causes a monitoring plot to fall outside the site
boundary, and then a later edit causes the plot to be part of the site again,
we should use it as a permanent plot before we resort to creating a brand-new
plot in the added area. This will increase the amount of observation data we
collect in the same locations over time.

Practically speaking, this means we no longer mark the plot as unavailable
when it's ejected. This is a change from the previous site editing behavior,
but previously we only needed to mark such plots as unavailable because they
were still associated with their original subzones.