SW-6405 Support flexible site edits in admin UI (#2806)
Add an option to the planting site shapefile upload form to allow selecting which
edit behavior to use. The new "Flexible" option is selected by default.

Show a more detailed preview of the edits that would be applied. This will be
critical for manual testing of the feature, but will also be useful for
sanity-checking real site changes.