Fixed schema gen issues and moved description and isDead to trunk (#2819)
`oneOf` dictates that one of the children is required be part of the payload. `discriminatorMapping` means that this can be polymorphed to one of the children. 

Subtle differences but it breaks down when the child component also uses `allOf` to inherit properties from the parent (which is the correct usage). `oneOf` in this case would lead to an infinite loop of dependencies. `discriminatorMapping` makes it all work correctly.