SW-6533 Remove redundant observation test
During the move from 25m to 30m plots, we added a copy of one of the observation
results tests where the plot size changed but everything else was the same, to
verify that only the expected stats would be affected by the size change. Now that
we've fully moved to 30m plots, the 25m test is unnecessary. Remove it so we don't
need to maintain it as the stats calculations change.