SW-6566 Support moving subzones between zones
Previously, if a subzone in a new shapefile had the same name as an existing
subzone but a different zone name, it was treated as a deletion of the existing
subzone and the creation of a new one. Add logic to detect this situation and
instead move the subzone from one zone to the other, preserving all its data such
as planting history and monitoring plots.

A subzone can move from one existing zone to another existing zone, or it can move
from an existing zone to a newly-created one.