SW-6561 Fix migration PostgreSQL 17 incompatibility
In PostgreSQL 17, `system_user` is now a reserved keyword, so we can't use it
unquoted as a subquery or CTE name any more.

We were using it in two migrations. One of them is a numbered migration, and
switching to a different CTE name means editing the migration script. By default,
Flyway flags edits to numbered migrations as validation errors. So we also need
to tell Flyway to not treat this specific edit as an error, but to still flag
edits to other numbered migrations.