SW-6653 Fix stack overflow on map update (#2881)
If a planting zone in an uploaded shapefile had an added area that shared a vertex
with a removed area, one of our geometry operations (calculating an intersection)
resulted in a geometry that contained a point as well as a set of polygons. This
caused infinite recursion in the code that converts arbitrary geometries to
multi-polygon objects.

Update the converter code to ignore standalone points and only pay attention to
polygons and multi-polygons.