SW-6520 Implemented system metric read structure (#2889)
This PR implements the query steps for reading system metric values and entries for a report. It defaults to frozen value if there is one, otherwise falls back to live system queried data. Currently, this is a placeholder number. 

The actual querying of Terraware data will follow in the next PRs.