SW-6524 Store method to refresh system metric values (#2901)
The refresh method will query Terraware system metric value, and write into the report_system_metrics table, effectively freezing the value. This method can be called by a report reviewer at any time, and also will be called when a report is submitted.

The API and the submission handler will be implemented in following PRs.