SW-6561 Use custom PostgreSQL 17 image for build/test
Make builds and tests use a custom version of PostgreSQL 17 that has the
extensions required by Terraware and is available for both ARM and x86.
This replaces the x86-only PostgreSQL 13 image we were using.

For now, we will continue to stick with PostgreSQL 13-compatible SQL since
the production environment hasn't been upgraded to 17 yet.