SW-6561 Use custom PostgreSQL 17 image for build/test (#2916)
Make builds and tests use a custom version of PostgreSQL 17 that has the
extensions required by Terraware and is available for both ARM and x86.
This replaces the x86-only PostgreSQL 13 image we were using.

For now, we will continue to stick with PostgreSQL 13-compatible SQL since
the production environment hasn't been upgraded to 17 yet.

The build/test configuration in `gradle.properties` uses a specific version
tag while the demo configuration in `docker/docker-compose.yml` uses a
generic `release` tag. The assumption is that the latter is used for quick
one-off runs of the server where you don't need fine control over the
exact version, whereas for builds, we want repeatability which means
we need to be specific about dependency versions.