SW-6587 Create Ask Terraware tables (#2928)
Create the database tables to support the experimental Ask Terraware feature,
which uses Spring AI to talk to the OpenAI chat API.

The big change here is the introduction of a table to store embeddings extracted
from uploaded documents and project details. It uses the pgvector extension to
support similarity searches on embeddings.