SW-6587 Ask Terraware implementation (#2932)
This experimental change adds integration with the OpenAI API to Terraware.

It is exposed via a page in the admin UI, `/admin/ask/`. Currently that page isn't
linked from elsewhere in the admin UI; we want to roll this feature out to a small
handful of specific internal users initially but we don't want to forbid others
from using it.

The chat UI uses the "htmx" library to do lightweight dynamic updates rather than
full page reloads.

Currently, projects must be "prepared" before you can ask the chat UI about them.
The "prepare" process generates embeddings for all the project's variable values
and uploaded documents as well as for some non-variable metadata like the project
accelerator details.

While we'll deploy this change to production for limited testing, it's still a
prototype; there are no tests and the technical documentation is incomplete.