SW-6696 Make subzone species list include entire site (#2934)
Currently, the mobile app's observation UI has a menu of known species, which is
based on the plants that are reported to have been planted in the monitoring
plot's subzone as well as plants that were previously observed at the site.

Now that subzone boundaries can be changed even after plants are planted, limiting
the menu to species in the plot's subzone no longer works, because the plot may be
in a different subzone now than it used to be.

Update the API endpoint that lists species in a subzone to include all species
planted at the site.