SW-6643 Updated workflow service to not copy unchanged workflow row (#2949)
This prevents a race condition from happening when a user updates the variable value as well as the variable workflow. This is common when reviewing documents/questionnaire deliverables when an accelerator admin may adjust a value but also change the status of the variable. This was causing an issue when the value update API call, takes the workflow row that is not the latest, and creates a new workflow history row with it. 

The result is something like this:
Newest:
Row 3, Old Status, New value ID (Created by the update value API call)
Row 2, New Status, Old value ID (Created by the update workflow API call)
Row 1, Old Status, Old value ID (Existing)

But we want a guarantee of this:
Row 3, New Status, New value ID

The update order doesn't actually matter that much.


Key changes:
~~1) Moved the fetch existing instead a transaction, and behind a select for update.~~
2) Change the update function to take an update function, instead of individual values. This is to simplify the call from `VariableValuesService` since it is also doing a fetch workflow there. 
3) The actual simplification part in the service class of step 2.