Return participant projects for species in order
`ParticipantProjectSpeciesStore.fetchParticipantProjectsForSpecies` was returning
results in an undefined order, but its tests were assuming they'd be in a specific
order. Usually that worked, but occasionally the tests would fail if the database
decided to order its query results differently.

Make the query return results in the order expected by the tests.