SW-6717 Don't double-count dead plants in moved plots (#2952)
If all the permanent monitoring plots from past observations are moved to a
different subzone by a map edit, we were counting their dead plants in both
their new subzone and the original one. Instead, we need to treat the original
subzone as if it's being observed for the first time and not try to carry the
cumulative dead plants total forward from past observations.