SW-6716 Zone/subzone names in observation results
Currently, observation results at the zone and subzone levels include zone and
subzone IDs. But if a map edit deletes an area, its ID will become invalid; we
will still have a record of it in the planting zone/subzone history tables,
but trying to search for it by ID won't work any more.

In preparation for updating the terraware-web code to properly display
observations of areas that were subsequently deleted, update the API to include
the names of zones and subzones alongside their IDs. The web app will use names
to identify zones and subzones in the context of a particular observation.

This is only a preparatory change; once terraware-web has been updated to use
these newly-added names a followup change will update the database queries
to read the zone/subzone names from the history tables and will update the API
to make the IDs nullable.