Updated refresh system metrics to set override value to null, and fill columns only when reports are submitted (#2971)
This is to change the function of resetting system metrics based on requirements.

The new requirement specifies that overwrites are always erased when refreshed. And this functionality will be used for when reports are not submitted as well.

Based on this requirement, the function now supports both behaviors, but correctly leaves the columns empty if a report has not yet been submitted.