SW-6716 Add areas to site history tables (#2969)
Some observation statistics are based on the areas of the site, zones, and
subzones, so we include the area of each region in observation results API
responses. Currently, the areas are based on the current boundaries of the
regions, but for sites whose maps were edited after an observation, the current
areas might not be the same as the areas at the time of observation.

Rather than calculating the areas on the fly at read time, record them at edit
time; the area for a given history entry will never change.

Add a migration to populate the values for existing history rows.