SW-6716 Use historical maps in observation results (#2970)
A planting site map edit can cause existing monitoring plots to move between
planting zones or subzones. Previously, the observation results fetching code only
looked at the _current_ location of each monitoring plot, so users looking at
the results of past observations, who expected to see the site map as it existed
at the time of each observation, would see monitoring plots listed under the wrong
zones. Worse, a monitoring plot in a zone that was subsequently deleted from the
site wouldn't show up in the observation details at all.

Change the results fetching logic to use the zone/subzone/plot associations from
the map that was active at the time of the observation.

If a plot was in a subzone that no longer exists, it won't have a subzone ID any
more, but we'll still be able to fetch the name of the subzone (and zone) that
existed during the observation. In that case, the subzone and zone IDs will be
null in the API response; update the response payload definitions accordingly.