Move insert*History defaults to parameters
The default values for some of the parameters to `insertPlantingSiteHistory` were
declared on the parameters themselves, but the defaults for the zone and subzone
history insert methods were calculated in the bodies of the methods. Move their
defaults to the parameter list for consistency and so that the parameters can be
non-nullable.