Fix intermittent PlantingZoneModelTest failures (#2981)
The assertion that was checking whether a monitoring plot was inside a specific
part of a planting zone map was too strict. We allow plots to fall _slightly_
outside the map boundaries to account for inaccuracy in the geometry calculations.

The test was failing if 0.001% or more of the plot was outside the requested
area, which was stricter than the fuzz factor we allow when searching for plot
locations. Increase the tolerance to 0.1%.

Also change the assertion to produce a more informative error message.