SW-6707 Fix species merging for ad-hoc observations
`/api/v1/tracking/observations/{observationId}/mergeOtherSpecies` is supposed to
work for both assigned observations and ad-hoc monitoring observations, but it was
broken in the ad-hoc case because when it was fetching the details of the
monitoring plot, it was calling a fetch function that was designed for assigned
plots that were in planting subzones.

Change it to query only the plot data it needs; this not only fixes the feature
for ad-hoc observations, but is slightly more efficient for assigned ones too.