SW-6694 Convert simple planting sites to detailed (#3011)
Add a manual migration to the admin UI to add zones and subzones to planting
sites that don't have them. This will allow users to schedule observations of
sites they created as simple planting sites.

There is still a way to create simple planting sites in the web app; we'll need
to run the migration again once that's removed, at which point the migration can
go away.