Reduce variable fetches (#3021)
Reduce calls for each of `projectAcceleratorVariablesStableIds` to one call to the VariableStore for all of them together. 

Future eventual enhancement: update `fetchByStableId` and new `fetchListByStableIds`  to add variables to the cache immediately instead of passing the ID to another fetch method for every value returned.

This extra fetching was noticed as part of SW-6786, since many more stable Ids will be added for that story.