SW-6189 Remove subplots from data model (#3027)
When permanent monitoring plots came in clusters of 4, we gave each plot in
the cluster a "subplot" number to differentiate them. Since we're no longer
doing multi-plot clusters, currently the subplot number is always 1 and
nothing reads it.

Drop the column from the database and remove references to it from the code.