SW-6189 Remove ObservationServiceTest cluster support
The test was already updated to only insert 1-plot clusters, but internally it
was still representing a permanent "cluster" as a list of plot IDs rather than
as a single ID. Simplify the implementation.

This doesn't remove the term "cluster" from the test code; that'll be done in a
subsequent change.