SW-6694 Backfill data from simple planting sites
Commit 0f7e1c7c added an admin function to convert simple planting sites to
detailed ones, but the migration only updated the site maps. It needs to also
migrate the withdrawal data, including the detailed records of plantings and
the species totals at the zone and subzone levels.

Migrate that data as well, in such a way that it will work for sites that have
already been converted by the previous version of the migration.