SW-6788 Add variableValueIds for project photos (#3034)
Return the variableValueIds in the project details endpoint. This will be used to call the endpoint `/api/v1/document-producer/projects/{projectId}/images/{valueId}` with the projectId and valueId to retrieve each photo.

A future PR (under the same story) will add an endpoint to generate a mapbox image on demand using the project's boundaries (from `planting_sites`).