SW-6766 Improve zone replacement/rename handling (#3046)
The flexible site editing code had support for zone boundaries changing such that
a subzone moved from one zone to another. However, it wasn't handling cases where
the original zone was completely deleted. Since we match the existing and new
zones using their names when a new shapefile is uploaded, this meant that we
weren't handling cases where a planting zone is renamed.

Update the edit calculator to properly move subzones to the new zones in that
case, and don't delete the subzones as part of the deletion of their old zones.