SW-6825 Add zone/subzone stable IDs (2/2)
Currently, when a new shapefile is uploaded for an existing planting site, we
match the zone and subzone names in the shapefile with the names in the
database to determine whether a zone/subzone is being created, deleted, or
modified.

Unfortunately, this means that we can't support renames very well: changing the
name causes the name-matching code to think that the existing zone/subzone is
being deleted and a new one is being created.

Address this by adding "stable ID" values to zones and subzones. This will work
similarly to the stable IDs of variables: it'll be used to determine the
identity of a zone/subzone.

The previous change in this series added the stable ID to the data model. This
change updates the planting site editing code to use it to determine
zone/subzone identity. It adds stricter validation of shapefiles to ensure
that site edits are unambiguous.