SW-6839 Disallow ad-hoc observations of simple sites (#3076)
Creating an observation of a planting site without a map currently fails with
an internal error because there's no planting site history ID to use when
inserting the ad-hoc plot.

Add a sanity check to make ad-hoc observations fail for simple planting sites
generally, not just ones without maps; the desired product behavior is to only
allow observations of detailed sites, whether the observations are assigned or
ad-hoc.