SW-6600 Generate embeddings for submitted documents
If a new document is submitted for a deliverable on a project that has been
prepared for Ask Terraware, automatically generate embeddings for the document
so its information will be available in Ask Terraware right away.

This is purely additive since there isn't currently a way for users to specify
that a newly-uploaded file replaces an existing one.