SW-6600 Regenerate embeddings for updated variables (#3079)
When a variable value is updated for a project that has been prepared for
Ask Terraware, generate new embeddings for just that variable. This will
keep Ask Terraware's responses up to date as project data is filled in
without the user having to manually re-prepare the project.

This only affects variable values. A separate change will add similar logic
for newly-submitted deliverable documents.