SW-6891 Allow contributors to mark subzones completed
Contributors and managers are supposed to be able to mark a subzone as "planting
completed" but the server was only letting admins and owners do it.

Internally, the completed flag was being set using an "update subzone" function,
but the function didn't actually allow editing any data other than planting
completion; the idea was that it could be extended to other edits later. But
that's likely to make permissions more complex since contributors probably won't
be allowed to edit other subzone data if/when we allow users to edit site maps.

Change the function and its associated permission to only refer to editing
subzones' completion status, and relax the permission to allow contributors and
managers to change the completion status.