SW-6890 Don't log error if module event is deleted (#3090)
If an admin deletes a module event, the expected behavior is that the scheduled
jobs for checking the event's status and sending notifications about it will see
that the event no longer exists and exit cleanly. We were logging error messages
in that case, even though it's not an error condition.

Update the log messages to "info" level and add an explanation of why they're
not errors.