SW-6896 Fix permission error on embedding update (#3093)
When a user without permission to read accelerator details updated a variable
value or uploaded a deliverable document, the event handler that enqueues a job
to update the project's embeddings was unable to check whether or not the
project had already had its embeddings generated. This didn't cause the original
operation to fail, but the embeddings never got updated.