SW-6895 Prune old timeseries values (#3094)
The `timeseries_values` table is by far the largest one in the production
database, and most of its contents are of no interest. We only need recent history
for most of the timeseries. However, for certain timeseries, we want to retain
historical data forever, so we can't just apply a fixed retention period across
the board.

Add a column to the timeseries table to configure how many days' worth of values
to retain, and add a recurring job that deletes old values based on each
timeseries' retention setting. To avoid accidentally deleting values that we need
to retain indefinitely, this uses an opt-in model: the default is to not prune any
old values, and we'll set the retention days on the specific timeseries we want to
trim. The opt-in approach also lets us roll this out gradually, starting with just
a few timeseries to get a better understanding of the real-world performance
impact.