SW-6912 Fixed duplicated report bugs (#3106)
This fixes an issue when updating report config dates, that multiple reports of the same dates get created. This was due to an incorrect assumption that reports are inserted/returned in order by the DAO.

This fixes it by sorting the results returned by the DAO, as well as removing all duplicated reports marked as "Not Needed".